% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{add_magnitude}
\alias{add_magnitude}
\title{Adds magnitude for each level of the doc structure}
\usage{
add_magnitude(doc_part)
}
\arguments{
\item{doc_part}{a structured list}
}
\value{
the structured list with added magnitude values
}
\description{
Adds magnitude for each level of the doc structure
}
