% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{df_to_krona}
\alias{df_to_krona}
\title{Convert a data.frame to a KRONA XML structure}
\usage{
df_to_krona(df, name, magnitude, ...)
}
\arguments{
\item{df}{the data.frame}

\item{name}{the name of the dataset}

\item{magnitude}{the name of the column which holds the magnitude values}

\item{...}{a list of columns on which to aggregate, note: order matters}
}
\value{
an xml object (based on xml2)
}
\description{
Convert a data.frame to a KRONA XML structure
}
