% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{df_to_krona_cached}
\alias{df_to_krona_cached}
\title{Cached version of df_to_krona}
\usage{
df_to_krona_cached(cache_type = "memory", cache_fs_path = NULL)
}
\arguments{
\item{cache_type}{either \code{memory} or \code{filesystem}}

\item{cache_fs_path}{directory to store cached data if \code{cache_type == "filesystem"}}
}
\value{
cached output of df_to_krona
}
\description{
Cached version of df_to_krona
}
