% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{do_rename}
\alias{do_rename}
\title{Rename key to 'node' and add original name as attribute}
\usage{
do_rename(doc_part)
}
\arguments{
\item{doc_part}{a structured list}
}
\value{
the structured list with renamed keys
}
\description{
Rename key to 'node' and add original name as attribute
}
