% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronawidget.R
\name{kronawidget}
\alias{kronawidget}
\title{<Add Title>}
\usage{
kronawidget(krona_df, width = "100\%", height = "500px", elementId = NULL)
}
\arguments{
\item{krona_df}{Dataframe}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Element ID}
}
\description{
<Add Description>
}
