% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transform.R
\name{save_krona_xml}
\alias{save_krona_xml}
\title{Save krona object}
\usage{
save_krona_xml(doc, filename)
}
\arguments{
\item{doc}{a KRONA document object created by \code{df_to_krona}}

\item{filename}{the filename to write the xml to}
}
\description{
Save krona object
}
