% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kronawidget.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{testdata}
\format{A test data frame to be used with krona_to_df
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\usage{
testdata
}
\description{
testdata
}
\keyword{datasets}
